+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - tests for task and actions creation
 - handle case of allOf in JSON schema when recursively retrieving "hai" print level
 - add message to task
+- fixed issue with type=array/items in JSON schema when recursively retrieving "hai" print level
 
 
 ## 0.2.8